Move SQLKit extension tests to own file
